Split fbgemm kernels into focused SYCL modules and decouple FbgemmOps
* Break out monolithic fbgemm kernel declarations/implementations into dedicated operator files:
** FbgemmCumsum
** JaggedTensorOps
** Permute2DSparseData
** ReorderBatchedAd
*Update FbgemmOps to include the new per-operator headers and remove dependency on the old shared kernel header.
* Delete legacy FbgemmKernels.h after moving reusable pieces to operator-local modules.
* Centralize scalar dispatch helpers in fbgemm_utils/dispatch_macros.h and reuse them from new modules.
* Keep behavior unchanged while aligning code structure with existing operator-specific split patterns used across the xpu SYCL backend.
