+++ src/ATen/native/xpu/sycl/FbgemmCumsum.h
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2020-2026 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+#pragma once
+
+#include <ATen/ATen.h>
+
+namespace at::native::xpu {
+
+TORCH_XPU_API void fbgemm_cumsum_kernel(
+    const Tensor& result,
+    const Tensor& self,
+    int64_t dim);
+
+} // namespace at::native::xpu
