+++ src/ATen/native/xpu/sycl/JaggedTensorOps.h
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2020-2026 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+#pragma once
+
+#include <ATen/ATen.h>
+
+namespace at::native::xpu {
+
+TORCH_XPU_API void dense_to_jagged_forward_xpu_kernel(
+    const Tensor& x_values,
+    const std::vector<Tensor>& x_offsets,
+    const Tensor& y,
+    const Tensor& output_values);
+
+TORCH_XPU_API void jagged_to_padded_dense_forward_xpu_kernel(
+    const Tensor& x_values,
+    const std::vector<Tensor>& x_offsets,
+    const Tensor& y,
+    const Tensor& output,
+    const double padding_value = 0.0);
+
+TORCH_XPU_API void jagged_dense_elementwise_add_jagged_output_fwd_xpu_kn(
+    const Tensor& x_values,
+    const std::vector<Tensor>& offsets,
+    const Tensor& dense,
+    const Tensor& output_values);
+
+} // namespace at::native::xpu
