+++ src/ATen/native/xpu/sycl/ReorderBatchedAd.h
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2020-2026 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+#pragma once
+
+#include <ATen/ATen.h>
+
+namespace at::native::xpu {
+
+TORCH_XPU_API void reorder_batched_ad_lengths_xpu_kernel(
+    const Tensor& cat_ad_lengths,
+    const Tensor& batch_offsets,
+    Tensor& reordered_cat_ad_lengths,
+    const int32_t T,
+    const bool broadcast_lengths,
+    const int32_t grid_size);
+
+TORCH_XPU_API void reorder_batched_ad_indices_xpu_kernel(
+    const at::Tensor& cat_ad_offsets,
+    const at::Tensor& cat_ad_indices,
+    const at::Tensor& reordered_cat_ad_offsets,
+    const at::Tensor& batch_offsets,
+    at::Tensor& reordered_cat_ad_indices,
+    const int64_t num_ads_in_batch,
+    const int64_t B,
+    const int64_t T,
+    const bool broadcast_indices = false);
+
+} // namespace at::native::xpu
